Assemble training data from aggregate sequence data

trainingData subsets the aggregate data of a ModifierML object to positions as defined by coord. Positions with an entry in the mod column are labeled TRUE.

trainingData(x, coord, ...)

# S4 method for ModifierML,GRanges
trainingData(x, coord, ...)

# S4 method for ModifierML,GRangesList
trainingData(x, coord, ...)

Arguments

x
a ModifierML object

coord
a GRanges or a GRangesList object

...
See subsetByCoord for more details. type is hard coded to FALSE to disregard subsetting by type. In addition, merge is set to FALSE by default, but it can be set to TRUE.

Value

a CompressedSplitDataFrameList with aggregate data and an addition label column.

See also

Examples

data("dmod",package = "RNAmodR.ML")
setClass("ModMLExample",
         contains = c("RNAModifierML"),
         prototype = list(mod = c("D"),
                          score = "score",
                          dataType = c("PileupSequenceData",
                                       "CoverageSequenceData"),
                          mlModel = character(0)))
data("me",package = "RNAmodR.ML")
nextUPos <- function(gr){
  nextU <- lapply(seq.int(1L,2L),
                  function(i){
                    subseq <- subseq(RNAmodR::sequences(me)[dmod$Parent], start(dmod)+3L)
                    pos <- start(dmod) + 2L +
                      vapply(strsplit(as.character(subseq),""),
                             function(y){which(y == "U")[i]},integer(1))
                    ans <- dmod
                    ranges(ans) <- IRanges(start = pos, width = 1L)
                    ans
                  })
  nextU <- do.call(c,nextU)
  nextU$mod <- NULL
  unique(nextU)
}
nondmod <- nextUPos(dmod)
nondmod <- nondmod[!(nondmod %in% dmod)]
coord <- unique(c(dmod,nondmod))
coord <- coord[order(as.integer(coord$Parent))]
trainingData(me,coord)

#> SplitDataFrameList of length 158
#> $`6`
#> DataFrame with 1 row and 17 columns
#>      d2.base d2.arrest   d1.base d1.arrest      base  coverage           G
#>    <numeric> <numeric> <numeric> <numeric> <numeric> <numeric>   <numeric>
#> 47         1  0.351759         1  0.997132         3     23129 6.66667e-05
#>             A         T          C    arrest   mismatch   u1.base  u1.arrest
#>     <numeric> <numeric>  <numeric> <numeric>  <numeric> <numeric>  <numeric>
#> 47 0.00187645  0.993217 0.00483961 0.0970643 0.00678273         4 0.00539715
#>      u2.base  u2.arrest    labels
#>    <numeric>  <numeric> <logical>
#> 47         3 0.00129261      TRUE
#> 
#> $`6`
#> DataFrame with 1 row and 17 columns
#>      d2.base  d2.arrest   d1.base d1.arrest      base  coverage          G
#>    <numeric>  <numeric> <numeric> <numeric> <numeric> <numeric>  <numeric>
#> 54         1 0.00772596         1 0.0127919         3     26397 0.00147485
#>              A         T          C     arrest   mismatch   u1.base  u1.arrest
#>      <numeric> <numeric>  <numeric>  <numeric>  <numeric> <numeric>  <numeric>
#> 54 3.33167e-05   0.99476 0.00373167 0.00483814 0.00523983         3 0.00706246
#>      u2.base  u2.arrest    labels
#>    <numeric>  <numeric> <logical>
#> 54         4 0.00818039     FALSE
#> 
#> $`6`
#> DataFrame with 1 row and 17 columns
#>      d2.base d2.arrest   d1.base  d1.arrest      base  coverage         G
#>    <numeric> <numeric> <numeric>  <numeric> <numeric> <numeric> <numeric>
#> 55         1 0.0127919         3 0.00483814         3   26525.3         0
#>              A         T          C     arrest   mismatch   u1.base  u1.arrest
#>      <numeric> <numeric>  <numeric>  <numeric>  <numeric> <numeric>  <numeric>
#> 55 6.66867e-05  0.997426 0.00250768 0.00706246 0.00257436         4 0.00818039
#>      u2.base u2.arrest    labels
#>    <numeric> <numeric> <logical>
#> 55         1 0.0206765     FALSE
#> 
#> ...
#> <155 more elements>